moves tiled camera cfg outside
