+++ source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/camera/__init__.py
@@ -6,7 +6,8 @@
 """Sub-module for camera wrapper around USD camera prim."""
 
 from .camera import Camera
-from .camera_cfg import CameraCfg, TiledCameraCfg
+from .camera_cfg import CameraCfg
 from .camera_data import CameraData
 from .tiled_camera import TiledCamera
+from .tiled_camera_cfg import TiledCameraCfg
 from .utils import *  # noqa: F401, F403
